code refactoring separated base class for entity separated entity and response-dto for Account

+++ src/main/java/com/spanprints/authservice/dto/inventory/AddToInventoryRequest.java
@@ -1,5 +1,7 @@
 package com.spanprints.authservice.dto.inventory;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,6 +17,6 @@ public class AddToInventoryRequest {
 	private Long itemId;
 	private Long count;
 	private Double rate;
-	private Double amountPaid;
+	private BigDecimal amountPaid;
 
 }
